+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Upcoming
 
+* Query validation rule: Unique variable names (#112)
 * Add suggested types to incorrect field message (#111)
 * Updated descriptions of a scalar values
 * Minor improvements